Manually location of the buildings creations
Now the player after clicking a building to construct has to click into
the map to position the building.
- Changed all Start() functions of buildings and units for the Awake().
Have problems with the create of buildings and the progressive
construction.
- Changed for units the function override createBuilding() by the
PerformAction(), now a unit can make different things than create a
Building only passing a String.
- Fixed the SelectedArea and created by script for the Userinput.cs and
removed from the map objects.
- Changed and optimized the CreateBuilding() function of CivilUnit.cs
- Changed the Update() of the CivilUnit.cs for the different states of
building